requirements.txt
django~=1.9.0
django-registration==2.2
django-multiselectfield==0.1.4